KOCHKA: add `en-US` stub file
This is a nice way how to deal with the default language (not to load special translations since there is nothing to translate).
